Add files via upload
An algorithm that produces all possible combinations of numbers from a list. Its strong point is that for each new set of combinations (i.e the set of combinations of 3 elements), it only looks at the previous set (i.e the set of combinations of 2 elements), thus avoiding useless computations. The trick is achieved by using a List of Lists and then flattening the List one level.